requirements.txt
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==5.3.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.8
Pillow==8.1.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
